finally fixed this asshole map rendering algorithm along with entities

+++ src/main/java/com/emkave/pacman/entity/Collectible.java
@@ -3,12 +3,11 @@ package com.emkave.pacman.entity;
 import javafx.scene.image.ImageView;
 
 public abstract class Collectible extends Entity { // Every collectible is an entity
-    Collectible(ImageView __imageView, final double x, final double y) {
+    Collectible(ImageView __imageView, final int x, final int y) {
         super(__imageView, x, y);
     }
 
 
     @Override public void render() {
-
     }
 }
